LayoutTests/svg/dynamic-updates/SVGEllipseElement-dom-cy-attr.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="resources/SVGTestCase.js"></script>
<script src="../../resources/js-test-pre.js"></script>
<script src="../../fast/repaint/resources/repaint.js"></script>
</head>
<body onload="runRepaintTest()">
<h1>SVG 1.1 dynamic update tests</h1>
<p id="description"></p>
<div id="console"></div>
<script>
// [Name] SVGEllipseElement-dom-cy-attr.js
// [Expected rendering result] unclipped green ellipse - and a series of PASS messages

description("Tests dynamic updates of the 'cy' attribute of the SVGEllipseElement object")
createSVGTestCase();

var ellipseElement = createSVGElement("ellipse");
ellipseElement.setAttribute("cx", "150");
ellipseElement.setAttribute("cy", "-50");
ellipseElement.setAttribute("rx", "100");
ellipseElement.setAttribute("ry", "150");
ellipseElement.setAttribute("fill", "green");

rootSVGElement.appendChild(ellipseElement);
shouldBeEqualToString("ellipseElement.getAttribute('cy')", "-50");

function repaintTest() {
    ellipseElement.setAttribute("cy", "150");
    shouldBeEqualToString("ellipseElement.getAttribute('cy')", "150");

    completeTest();
}

var successfullyParsed = true;
</script>
</body>
</html>
