third_party/blink/web_tests/svg/dynamic-updates/SVGEllipseElement-dom-cy-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'cy' attribute of the SVGEllipseElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ellipseElement.getAttribute('cy') is "-50"
PASS ellipseElement.getAttribute('cy') is "150"
PASS successfullyParsed is true

TEST COMPLETE
